(PC-8434) Filter hits without subcategoryId from Algolia/AppSearch
Due to late indexation of field subcategoryId, some documents do not
have this field, especially on Algolia (ex: objectId 40488429).
Since the frontend expects this to be defined, it crashes at some point.
As a solution, we filter hits from both search backend that do not have
this field (post search). This can be reverted once the migration to
AppSearch is complete because all documents on AppSearch have this field.