(PC-18595) feat(useRecommendedHits): get  eventDuringNextXDays from contentful and compute beginning and ending datetime for each playlist
The test have been modified to only use date with no timezone (UTC). At the moment, the native tests are launched with the TZ=UTC option and therefore, all the new Date() are created with the UTC timezone. To be consistent and avoid breaking the existing tests, I changed the stringify dates by UTC ones