(PC-18595) feat(RecommendedHits): create a Contentful filter X days (#3950)
* (PC-18595) feat(homePlaylist): create a computeBeginningAndEndingDateTime util function in order to compute beginning and ending datetime of a playlist depending on contentful outputs

* (PC-18595) feat(contentful): add eventDuringNextXDays field to the RecommendationParametersFields contentful type

* (PC-18595) feat(useRecommendedHits): get  eventDuringNextXDays from contentful and compute beginning and ending datetime for each playlist

The test have been modified to only use date with no timezone (UTC). At the moment, the native tests are launched with the TZ=UTC option and therefore, all the new Date() are created with the UTC timezone. To be consistent and avoid breaking the existing tests, I changed the stringify dates by UTC ones

* (PC-18595) feat(homeAPI): move tests next to their component/hook