(PC-20926) feat(useStepperInfo): remove condition on ValidationPhoneStep
On the former logic of useSubscriptionSteps, the back was sending the nextSubscriptionSteps of a user. If the nextSubscriptionSteps included phoneValidation, the front computed a list of 4 buttons, if not the list only contained 3 buttons. In the new logic, the 4 buttons are computed and the front will decide lately which one to disply or not