(PC-21060) feat(Stepper): design error step (#4596)
* (PC-21060) refactor(IconDone): remove unused Icons

* (PC-21060) feat(TryAgain): add a TryAgain icon

* (PC-21060) feat(IconRetryStep): create a generic IconRetryStep component

This component is based on IconStepDone. Nowadays, the only retryable step is the identification one but later it could be extend to other steps.

* (PC-21060) chore(stepperFixture): rename fixture

* (PC-21060) feat(StepButtonState): add a retry state to the StepButtonState type

In this commit I only add a retry icon to the StepButtonState but I don't introduce the associated stepButton. For now, a Retryable step will be displayed as a current step. In the next commit the retryable step button will be introduced

* (PC-21060) feat(StepButton): create a retryable stepButton